Project 4 Part 1 with Waspmote v15 and Smart Cities PRO
When ultrasound sensor detects a person passing through, a green LED will be turned on. Temperature will be taken into account as well. When an 'high' temperature has been reached, a red LED will be turned on, turning out the green LED in the case it was turned on. On the other hand, luminosity meassurement will be screened every 20 seconds, showing a message indicating if it's day or night.

In addition, any time, whereas the button has been pressed, the green led will be turned on for 10 seconds.